I changed the interval parameter to receive null inputs. For some reason, errors were appearing and the fix was to use a static cast on the interval parameter.

+++ Gordon360/Services/HousingService.cs
@@ -1208,7 +1208,7 @@ public class HousingService(CCTContext context) : IHousingService
             StatusName = newStatus.Status_Name,
             IsRecurring = newStatus.Is_Recurring,
             Frequency = newStatus.Frequency,
-            Interval = newStatus.Interval,
+            Interval = (int)newStatus.Interval,
             StartDate = newStatus.Start_Date,
             EndDate = newStatus.End_Date,
             CreatedDate = newStatus.Created_Date
